+++ server/controllers/teamController.js
@@ -45,12 +45,17 @@ module.exports.getTeam = function(req, res, next) {
 module.exports.postTeams = function(req, res, next) {
 
   let team_name = req.body.team_name,
+      team_type = req.body.team_type,
+      color = req.body.color,
+      goal = req.body.goal,
+      about = req.body.about,
       address1 = req.body.address1,
       address2 = req.body.address2,
+      team_city = req.body.team_city,
       team_state = req.body.team_state,
       zipcode = req.body.zipcode,
       country = req.body.country,
-      about = req.body.about,
+      username = req.body.username,
       leader = req.body.leader;
 
   //validate team inputs here
@@ -58,12 +63,17 @@ module.exports.postTeams = function(req, res, next) {
 
   let newTeam = {
     team_name: team_name,
+    team_type: team_type,
+    color: color,
+    goal: goal,
+    about: about,
     address1: address1,
     address2: address2,
+    team_city: team_city,
     team_state: team_state,
     zipcode: zipcode,
     country: country,
-    about: about,
+    username: username,
     leader: leader
 
   }
@@ -80,6 +90,14 @@ module.exports.postTeams = function(req, res, next) {
       //Update User's Team and User's Admin Level
       console.log("Team info from DB: " + team.uuid + ", team name: " + team.team_name);
 
+
+        Model.User.update(
+        {
+          team_id: team.uuid,
+        },
+        {
+          where: { username: team.leader }
+        })
       // userController.updateUserTeam(team);
 
      }).catch(err => {
